libbpf: ringbuf: allow to partially consume items
Instead of always consuming all items from a ring buffer in a greedy
way, allow to stop when the callback returns a value > 0.

This allows to distinguish between an error condition and an intentional
stop condition by returning a non-negative, non-zero value from the ring
buffer callback.

This can be useful for example to consume only a single item from the
ring buffer, every time that consume() or poll() are used.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>