libbpf: ringbuf: allow to consume up to a certain amount of items
In some cases instead of always consuming all items from a ring buffer
in a greedy way, we may want to consume up to a certain amount of items,
for example when we need to copy items from the BPF ring buffer to a
limited user buffer.

This allows to consume up to INT_MAX items from one (or more) ring
buffers without any risk to lose items.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>